django-backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-environ>=0.11,<1.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
boto3>=1.34,<2.0
stripe>=9.0,<10.0
gunicorn>=22.0,<23.0
ruff>=0.5.0,<1.0
flake8>=7.0,<8.0
